+++ integration_tests/suite/helpers/helpers/__init__.py
@@ -107,6 +107,7 @@ import entity
 import extension
 import funckey_template
 import group
+import group_call_permission
 import group_extension
 import group_member_extension
 import group_member_user